+++ fbpcs/private_computation/service/compute_metrics_stage_service.py
@@ -190,6 +190,10 @@ class ComputeMetricsStageService(PrivateComputationStageService):
             "concurrency": private_computation_instance.infra_config.mpc_compute_concurrency,
             "run_id": private_computation_instance.infra_config.run_id,
         }
+        if private_computation_instance.feature_flags is not None:
+            common_compute_game_args[
+                "pc_feature_flags"
+            ] = private_computation_instance.feature_flags
 
         game_args = []