+++ fbpcs/private_computation/service/pcf2_lift_stage_service.py
@@ -184,6 +184,10 @@ class PCF2LiftStageService(PrivateComputationStageService):
             "log_cost": self._log_cost_to_s3,
             "run_id": private_computation_instance.infra_config.run_id,
         }
+        if private_computation_instance.feature_flags is not None:
+            common_compute_game_args[
+                "pc_feature_flags"
+            ] = private_computation_instance.feature_flags
 
         game_args = []